requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
opencage==1.2
pylint==2.4.4
python-dotenv==0.11.0
requests==2.23.0
six==1.14.0
tornado==6.0.3
typed-ast==1.4.1
urllib3==1.25.8
websocket-client==0.57.0
wrapt==1.11.2
